Refactor roster page into modular components
